fix: fixed "no map detected" state on closing map & ftp exception
There is no internal state for if the map is
open or closed. There is a map detected
in a screenshot or not, that is the most
reliable way to find that out (afaik).

But it is misleading to users, if they open the map, the screenshot is uploaded and they close the map immediatly to see:
"No map detected" when there was a
map captured and uploaded, before they
closed it.

Now the state shows, that the map was probably closed, but
the last upload was successful.

Also, FTP exceptions (which these FluentFTP methods shouldn't
even throw) are catched and details are logged and
I foud out about String.IsBlank().