fix: now differentiates between extra mouse buttons 1 & 2 + capture task in new thread
The HotKey script does not block anywhere anymore, when a capture process is running. It starts a thread with the capture process instead of waiting on in it's own thread. There is also a check if the capture process thread is running, so that only one capture process is running at all times.