fix: HLL update broke detection on 1080p
Black Matter decided to shrink the map screen, making the image more
blurry. With that the maps bottom border became too slim to detect.
The fix is a 1080p specific detection which depends on the exact position
and size of the map. This will work even if the map is made bigger again,
but will break once the map is shrunk again in a HLL update.